+++ NEWS.md
@@ -1,3 +1,10 @@
+# reactR 0.1.2
+
+* Update to react `15.5.0`
+* Update to babel `6.24.0`
+* Add [office-fabric-react](http://dev.office.com/fabric) example
+* Build doc site with `pkgdown`
+
 # reactR 0.1.1
 
 * Added a `NEWS.md` file to track changes to the package.